Rename to TracerProviderBuilderExtensions
