+++ src/OpenTelemetry.Exporter.Zipkin/TracerProviderBuilderExtensions.cs
@@ -1,4 +1,4 @@
-﻿// <copyright file="OpenTelemetryBuilderExtensions.cs" company="OpenTelemetry Authors">
+﻿// <copyright file="TracerProviderBuilderExtensions.cs" company="OpenTelemetry Authors">
 // Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,20 +15,23 @@
 // </copyright>
 
 using System;
-using OpenTelemetry.Exporter.Console;
+using OpenTelemetry.Exporter.Zipkin;
 
 namespace OpenTelemetry.Trace
 {
-    public static class OpenTelemetryBuilderExtensions
+    /// <summary>
+    /// Extension methods to simplify registering of Zipkin exporter.
+    /// </summary>
+    public static class TracerProviderBuilderExtensions
     {
         /// <summary>
-        /// Registers a ConsoleActivity exporter to a processing pipeline.
+        /// Registers a Zipkin exporter that will receive <see cref="System.Diagnostics.Activity"/> instances.
         /// </summary>
         /// <param name="builder"><see cref="TracerProviderBuilder"/> builder to use.</param>
         /// <param name="configure">Exporter configuration options.</param>
         /// <param name="processorConfigure">Activity processor configuration.</param>
         /// <returns>The instance of <see cref="TracerProviderBuilder"/> to chain the calls.</returns>
-        public static TracerProviderBuilder UseConsoleExporter(this TracerProviderBuilder builder, Action<ConsoleExporterOptions> configure = null, Action<ActivityProcessorPipelineBuilder> processorConfigure = null)
+        public static TracerProviderBuilder UseZipkinExporter(this TracerProviderBuilder builder, Action<ZipkinExporterOptions> configure = null, Action<ActivityProcessorPipelineBuilder> processorConfigure = null)
         {
             if (builder == null)
             {
@@ -37,12 +40,12 @@ namespace OpenTelemetry.Trace
 
             return builder.AddProcessorPipeline(pipeline =>
             {
-                var exporterOptions = new ConsoleExporterOptions();
-                configure?.Invoke(exporterOptions);
+                var options = new ZipkinExporterOptions();
+                configure?.Invoke(options);
 
-                var consoleExporter = new ConsoleExporter(exporterOptions);
+                var exporter = new ZipkinExporter(options);
                 processorConfigure?.Invoke(pipeline);
-                pipeline.SetExporter(consoleExporter);
+                pipeline.SetExporter(exporter);
             });
         }
     }
